+++ src/loggers/index.d.ts
@@ -10,6 +10,7 @@ import File = require("./file");
 import Log4js = require("./log4js");
 import Pino = require("./pino");
 import Winston = require("./winston");
+import { LoggerOptions } from "./base";
 
 export {
 	Logger as Base,
@@ -27,5 +28,5 @@ export {
 
 export type { LogLevels } from "./base";
 
-export declare function resolve(opt: Record<string, any> | string): Logger;
-export declare function register(name: string, value: Logger): void;
+export declare function resolve(opt: Record<string, any> | string): Logger<any>;
+export declare function register(name: string, value: Logger<any>): void;